requirements.txt
aiofiles==23.1.0
aiosqlite==0.19.0
python-dotenv==1.0.0
python-telegram-bot==20.4
qrcode==7.4.2
jinja2

